+++ echopype/convert/parsed_to_zarr.py
@@ -8,8 +8,7 @@ import numpy as np
 import pandas as pd
 import zarr
 
-from ..core import ECHOPYPE_DIR
-from ..utils.io import check_file_permissions
+from ..utils.io import ECHOPYPE_DIR, check_file_permissions
 
 
 class Parsed2Zarr:
