+++ echopype/core.py
@@ -24,14 +24,6 @@ if TYPE_CHECKING:
     FileFormatHint = Literal[".nc", ".zarr"]
     EngineHint = Literal["netcdf4", "zarr"]
 
-ECHOPYPE_DIR = Path(os.path.expanduser("~")) / ".echopype"
-
-
-def init_ep_dir():
-    """Initialize hidden directory for echopype"""
-    if not ECHOPYPE_DIR.exists():
-        ECHOPYPE_DIR.mkdir(exist_ok=True)
-
 
 def validate_azfp_ext(test_ext: str):
     if not re.fullmatch(r"\.\d{2}[a-zA-Z]", test_ext):
